Properly kill sandbox systemd units when stage-1 is killed
This is way more complicated code than I'd like. An alternative is to
kill the bwrap process on timeout (which we can do -- the stage-2
process that _starts_ bwrap is still trusted code and can thus give us
its PID), but that's unreliable because we can only kill that based on
PID, and PIDs may be reused (unlikely, but possible).